Move plugins to @markbind/core
