+++ packages/core/package.json
@@ -21,8 +21,10 @@
     "@sindresorhus/slugify": "^0.9.1",
     "bluebird": "^3.7.2",
     "cheerio": "^0.22.0",
+    "crypto-js": "^4.0.0",
     "ensure-posix-path": "^1.1.1",
     "fastmatter": "^2.1.1",
+    "fs-extra-promise": "^1.0.1",
     "highlight.js": "^9.14.2",
     "htmlparser2": "^3.10.1",
     "lodash": "^4.17.15",
