+++ packages/core/src/constants.js
@@ -6,4 +6,18 @@ module.exports = {
 
   // packages/core/src/utils.js
   markdownFileExts: ['md', 'mbd', 'mbdf'],
+
+  // packages/core/src/util/fsUtil.js
+  sourceFileExtNames: ['.html', '.md', '.mbd', '.mbdf'],
+
+  // packages/core/src/plugins/algolia.js
+  ALGOLIA_CSS_URL: 'https://cdn.jsdelivr.net/npm/docsearch.js@2/dist/cdn/docsearch.min.css',
+  ALGOLIA_JS_URL: 'https://cdn.jsdelivr.net/npm/docsearch.js@2/dist/cdn/docsearch.min.js',
+  ALGOLIA_INPUT_SELECTOR: '#algolia-search-input',
+
+  // packages/core/src/plugins/default/markbind-plugin-plantuml.js
+  ERR_PROCESSING: 'Error processing',
+
+  // packages/core/src/plugins/utils/index.js
+  ERR_READING: 'Error reading',
 };
