+++ packages/core/src/plugins/algolia.js
