+++ packages/core/src/plugins/default/markbind-plugin-anchors.js
