+++ packages/core/src/plugins/default/markbind-plugin-shorthandSyntax.js
