+++ packages/core/src/plugins/filterTags.js
