+++ packages/core/src/plugins/googleAnalytics.js
