+++ packages/core/src/utils/fsUtil.js
